회원가입 및 자기정보 호출

+++ src/main/java/project/financial/domain/User.java
@@ -35,7 +35,7 @@ public class User implements UserDetails {
     @Column(nullable = true, length = 100)
     private String name;
 
-    private Long account;
+    private Integer account;
     private LocalDateTime dateTime;
 
     @ElementCollection(fetch = FetchType.EAGER)
